CTX-4879: Discussion fixes on PR.
