+++ coretex/cli/docker.py
@@ -1,4 +1,4 @@
-import subprocess
+from ..utils import command
 
 
 class DockerException(Exception):
@@ -11,33 +11,27 @@ class DockerException(Exception):
 
 
 def createNetwork(name: str) -> None:
-    try:
-        subprocess.run(["docker", "network", "create", "--driver", "bridge", name], check=True)
-    except:
-        raise DockerException(f"Failed to create network: {name}")
+    command(["docker", "network", "create", "--driver", "bridge", name])
 
 
 def removeNetwork(name: str) -> None:
-    try:
-        subprocess.run(["docker", "network", "rm", name], check = True)
-    except:
-        raise DockerException(f"Failed to remove network: {name}")
+    command(["docker", "network", "rm", name])
 
 
 def runNode(
-        name: str,
-        dockerImage: str,
-        imageType: str,
-        serverUrl: str,
-        storagePath: str,
-        nodeAccessToken: str,
-        nodeRam: str,
-        nodeSwap: str,
-        nodeSharedMemory: str,
-        restartPolicy: str,
-        ports: str,
-        capAdd: str,
-    ) -> bool:
+    name: str,
+    dockerImage: str,
+    imageType: str,
+    serverUrl: str,
+    storagePath: str,
+    nodeAccessToken: str,
+    nodeRam: str,
+    nodeSwap: str,
+    nodeSharedMemory: str,
+    restartPolicy: str,
+    ports: str,
+    capAdd: str,
+) -> None:
 
     runCommand = [
         "docker", "run", "-d",
@@ -58,35 +52,14 @@ def runNode(
         runCommand.extend(["--gpus", "all"])
 
     runCommand.append(dockerImage)
-
-    subprocess.run(runCommand, check=True)
-
-    # Check if the container is running
-    containerRunning = subprocess.run(["docker", "inspect", "--format", "{{.State.Running}}", name], capture_output=True, text=True).stdout.strip()
-
-    if containerRunning == "true":
-        return True
-
-    return False
+    command(runCommand)
 
 
 def stopContainer(name: str) -> None:
-    try:
-        # Disconnect and stop the container
-        subprocess.run(["docker", "stop", name], check=True)
-
-        # Remove the container
-        subprocess.run(["docker", "rm", name], check=True)
-    except:
-        raise DockerException(f"Failed to stop container {name}.")
+    command(["docker", "stop", name])
+    command(["docker", "rm", name])
 
 
-def stopNode(name: str, networkName: str) -> bool:
-    try:
-        stopContainer(name)
-
-        removeNetwork(networkName)
-
-        return True
-    except:
-        raise DockerException(f"Error occurred while stopping container {name}.")
+def stopNode(name: str, networkName: str) -> None:
+    stopContainer(name)
+    removeNetwork(networkName)
