+++ coretex/utils/__init__.py
@@ -20,3 +20,4 @@ from .file import guessMimeType, InvalidFileExtension
 from .date import DATE_FORMAT, TIME_ZONE, decodeDate
 from .hash import hashCacheName
 from .image import resizeWithPadding
+from .process import logProcessOutput, command, CommandException
